[alloc_policy]: Move AllocPolicy enum to the root
* Moved the AllocPolicy enum and its implementation to the root of
  crate as it will be used by the AddressAllocator and IdAllocator
  as well.
