+++ src/allocation_engine/interval_tree.rs
@@ -2,27 +2,8 @@
 // Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2
 
-use crate::{Error, Range, Result};
+use crate::{AllocPolicy, Error, Range, Result};
 use std::cmp::{max, Ordering};
-
-/// Policy for resource allocation.
-#[derive(Copy, Clone, Debug, PartialEq)]
-#[allow(dead_code)]
-pub enum AllocPolicy {
-    /// Allocate the first matched entry.
-    FirstMatch,
-    /// Allocate first matched entry from the end of the range.
-    LastMatch,
-    /// Allocate a memory slot starting with the specified address
-    /// if it is available.
-    ExactMatch(u64),
-}
-
-impl Default for AllocPolicy {
-    fn default() -> Self {
-        AllocPolicy::FirstMatch
-    }
-}
 
 /// Struct to describe resource allocation constraints.
 #[derive(Copy, Clone, Debug, PartialEq)]
