+++ src/lib.rs
@@ -99,6 +99,24 @@ impl Range {
     }
 }
 
+/// Policy for resource allocation.
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub enum AllocPolicy {
+    /// Allocate the first matched entry.
+    FirstMatch,
+    /// Allocate first matched entry from the end of the range.
+    LastMatch,
+    /// Allocate a memory slot starting with the specified address
+    /// if it is available.
+    ExactMatch(u64),
+}
+
+impl Default for AllocPolicy {
+    fn default() -> Self {
+        AllocPolicy::FirstMatch
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
